pyproject.toml
[tool.poetry]
name = "chatgpt_cli"
version = "0.1.0"
description = "ChatGPT from CLI"
authors = ["narita-atsuhiro <alayashiki@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
prompt-toolkit = "^3.0.38"
pyyaml = "^6.0"
rich = "^13.3.2"
fire = "^0.5.0"
requests = "^2.28.2"
dataclasses-json = "^0.5.7"
halo = "^0.0.31"
pydantic = "^1.10.7"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
mypy = "^1.1.1"

[tool.poetry.scripts]
chatgpt_cli = "chatgpt_cli.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
